Refactored some train.py functions to be in a utils file
